data-ingestion/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.*
requests==2.* 